chore: replace request access related pages
