+++ client/app/(authenticated)/dashboard/select-operator/request-access/confirm/[id]/page.tsx
@@ -1,0 +1,11 @@
+import { Suspense } from "react";
+import Loading from "@/app/components/loading/SkeletonGrid";
+import ConfirmSelectedOperator from "@/app/components/routes/select-operator/form/ConfirmSelectedOperator";
+
+export default async function Page({ params }: { params: { id: number } }) {
+  return (
+    <Suspense fallback={<Loading />}>
+      <ConfirmSelectedOperator params={params} />
+    </Suspense>
+  );
+}
