+++ client/app/(authenticated)/dashboard/select-operator/request-access/received/[id]/page.tsx
@@ -1,0 +1,11 @@
+import { Suspense } from "react";
+import Loading from "@/app/components/loading/SkeletonGrid";
+import AccessRequestReceived from "@/app/components/routes/select-operator/form/AccessRequestReceived";
+
+export default async function Page({ params }: { params: { id: number } }) {
+  return (
+    <Suspense fallback={<Loading />}>
+      <AccessRequestReceived params={params} />
+    </Suspense>
+  );
+}
