+++ client/app/(authenticated)/dashboard/select-operator/request-access/user-operator/[id]/page.tsx
@@ -1,0 +1,11 @@
+import { Suspense } from "react";
+import Loading from "@/app/components/loading/SkeletonGrid";
+import UserOperator from "@/app/components/routes/select-operator/form/UserOperator";
+
+export default async function Page({ params }: { params: { id: number } }) {
+  return (
+    <Suspense fallback={<Loading />}>
+      <UserOperator params={params} />
+    </Suspense>
+  );
+}
