--- client/app/select-operator/page.tsx
@@ -1,16 +1,0 @@
-import { Suspense } from "react";
-import Loading from "@/app/components/loading/SkeletonGrid";
-import PageBar from "@/app/components/PageBar";
-import SelectOperator from "@/app/components/routes/select-operator/form/SelectOperator";
-
-export default async function Page() {
-  const pageBarLabel: JSX.Element = <div>Select Operator</div>;
-  return (
-    <>
-      <PageBar label={pageBarLabel} />
-      <Suspense fallback={<Loading />}>
-        <SelectOperator />
-      </Suspense>
-    </>
-  );
-}
